aix: split code, that uses perfstat, to _cgo and _nocgo files
