+++ load/load_aix_nocgo.go
@@ -1,0 +1,16 @@
+//go:build aix && !cgo
+// +build aix,!cgo
+
+package load
+
+import (
+	"context"
+)
+
+func AvgWithContext(ctx context.Context) (*AvgStat, error) {
+	return nil, common.ErrNotImplementedError
+}
+
+func MiscWithContext(ctx context.Context) (*MiscStat, error) {
+	return nil, common.ErrNotImplementedError
+}
