+++ mem/mem_aix_nocgo.go
@@ -1,0 +1,16 @@
+//go:build aix && !cgo
+// +build aix,!cgo
+
+package mem
+
+import (
+	"context"
+)
+
+func VirtualMemoryWithContext(ctx context.Context) (*VirtualMemoryStat, error) {
+	return nil, common.ErrNotImplementedError
+}
+
+func SwapMemoryWithContext(ctx context.Context) (*SwapMemoryStat, error) {
+	return nil, common.ErrNotImplementedError
+}
